requirements.txt
python-telegram-bot==20.7
aiosqlite
pip install -r requirements.txt
